Code review quick wins: rename _special_forms, fix pyclass module, pin nohash-hasher, add module docs and docstrings
