+++ .gitignore
@@ -72,3 +72,7 @@ docs/_build/
 
 # Pyenv
 .python-version
+
+# PGO artifacts
+*.profdata
+*.profraw
