+++ tests/test_special_forms_py312.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 
 from serpyco_rs._meta import Annotations
-from serpyco_rs._secial_forms import unwrap_special_forms
+from serpyco_rs._special_forms import unwrap_special_forms
 from serpyco_rs.metadata import Min
